Refactor the persistance project
